some more .def support and various small fixes
git-svn-id: http://www.linux-addicted.net/repos/tinns@123 ef8e7eb3-8d0b-0410-9eab-9562649fb893

+++ server/src/game/include/def_shots.h
@@ -37,7 +37,7 @@ class PDefShot : public PDef
     int mDamageId;
     int mMass;
     int mRadius;
-    int mSpeed;
+    float mSpeed;
     //int mHitSound;
     //int mCustomType;
     //std::string mCustomFrameFx;
